block and file plugin related code seperation
