+++ csi/deploy/kubernetes/csi-attacher-opensdsplugin.yaml
