+++ csi/deploy/kubernetes/csi-nodeplugin-opensdsplugin.yaml
