+++ csi/examples/kubernetes/CreateVolumefromSnapshot/nginx.yaml
