+++ csi/examples/kubernetes/CreateVolumefromSnapshot/pvc_sc.yaml
