+++ csi/examples/kubernetes/CreateVolumefromSnapshot/restored-nginx.yaml
