+++ csi/examples/kubernetes/CreateVolumefromSnapshot/snapshot.yaml
