+++ csi/examples/kubernetes/fileshare.yaml
