+++ csi/examples/kubernetes/nginx.yaml
