+++ csi/examples/kubernetes/pod-with-block-volume.yaml
