+++ csi/plugins/block/identity.go
@@ -12,17 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package opensds
+package block
 
 import (
-	"fmt"
+	"github.com/opensds/nbp/csi/common"
 
 	csi "github.com/container-storage-interface/spec/lib/go/csi"
-	"github.com/golang/glog"
-	sdscontroller "github.com/opensds/nbp/client/opensds"
 	"golang.org/x/net/context"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 )
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -35,17 +31,7 @@ func (p *Plugin) Probe(
 	req *csi.ProbeRequest) (
 	*csi.ProbeResponse, error) {
 
-	glog.Info("start to probe")
-	defer glog.Info("end to probe")
-
-	_, err := sdscontroller.GetClient("", "")
-	if err != nil {
-		msg := fmt.Sprintf("failed to communicate with opensds client, %v", err)
-		glog.Error(msg)
-		return nil, status.Error(codes.FailedPrecondition, msg)
-	}
-
-	return &csi.ProbeResponse{}, nil
+	return common.Probe(ctx, req)
 }
 
 // GetPluginInfo implementation
@@ -54,13 +40,7 @@ func (p *Plugin) GetPluginInfo(
 	req *csi.GetPluginInfoRequest) (
 	*csi.GetPluginInfoResponse, error) {
 
-	glog.Info("start to get plugin info")
-	defer glog.Info("end to get plugin info")
-
-	return &csi.GetPluginInfoResponse{
-		Name:          PluginName,
-		VendorVersion: "",
-	}, nil
+	return common.GetPluginInfo(ctx, req, PluginName)
 }
 
 // GetPluginCapabilities implementation
@@ -69,22 +49,5 @@ func (p *Plugin) GetPluginCapabilities(
 	req *csi.GetPluginCapabilitiesRequest) (
 	*csi.GetPluginCapabilitiesResponse, error) {
 
-	return &csi.GetPluginCapabilitiesResponse{
-		Capabilities: []*csi.PluginCapability{
-			{
-				Type: &csi.PluginCapability_Service_{
-					Service: &csi.PluginCapability_Service{
-						Type: csi.PluginCapability_Service_CONTROLLER_SERVICE,
-					},
-				},
-			},
-			{
-				Type: &csi.PluginCapability_Service_{
-					Service: &csi.PluginCapability_Service{
-						Type: csi.PluginCapability_Service_VOLUME_ACCESSIBILITY_CONSTRAINTS,
-					},
-				},
-			},
-		},
-	}, nil
+	return common.GetPluginCapabilities(ctx, req)
 }
