+++ csi/plugins/block/plugin.go
@@ -12,15 +12,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package plugin
+package block
 
 import (
-	csi "github.com/container-storage-interface/spec/lib/go/csi"
+	"github.com/opensds/nbp/csi/common"
+	"github.com/opensds/opensds/client"
 )
 
-// Service Define CSI Interface
-type Service interface {
-	csi.IdentityServer
-	csi.ControllerServer
-	csi.NodeServer
+// Plugin define
+type Plugin struct {
+	VolumeClient *Volume
+}
+
+// NewServer Initialises and return plugin server
+func NewServer(client *client.Client) (common.Service, error) {
+	p := &Plugin{
+		VolumeClient: NewVolume(client),
+	}
+
+	// When there are multiple volumes unmount at the same time,
+	// it will cause conflicts related to the state machine,
+	// so start a watch list to let the volumes unmount one by one.
+	go common.UnpublishRoutine(client)
+
+	return p, nil
 }
