+++ csi/plugins/file/constants.go
@@ -1,0 +1,35 @@
+// Copyright 2018 The OpenSDS Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package file
+
+// fileshare constant parameters
+const (
+	ShareName       = "shareName"
+	ShareAZ         = "shareAZ"
+	ShareStatus     = "shareStatus"
+	SharePoolId     = "sharePoolId"
+	ShareProfileId  = "shareProfileId"
+	ShareProtocol   = "shareProtocol"
+	NFS             = "nfs"
+	ExportLocations = "exportLocations"
+	FileShareName   = "fileShareName"
+)
+
+// PluginName setting
+const (
+	PluginName = "csi-opensdsplugin-file"
+)
+
+var TopologyZoneKey = "topology." + PluginName + "/zone"
