+++ csi/plugins/file/identity.go
@@ -1,0 +1,53 @@
+// Copyright 2018 The OpenSDS Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package file
+
+import (
+	"github.com/opensds/nbp/csi/common"
+
+	csi "github.com/container-storage-interface/spec/lib/go/csi"
+	"golang.org/x/net/context"
+)
+
+////////////////////////////////////////////////////////////////////////////////
+//                            Identity Service                                //
+////////////////////////////////////////////////////////////////////////////////
+
+// Probe implementation
+func (p *Plugin) Probe(
+	ctx context.Context,
+	req *csi.ProbeRequest) (
+	*csi.ProbeResponse, error) {
+
+	return common.Probe(ctx, req)
+}
+
+// GetPluginInfo implementation
+func (p *Plugin) GetPluginInfo(
+	ctx context.Context,
+	req *csi.GetPluginInfoRequest) (
+	*csi.GetPluginInfoResponse, error) {
+
+	return common.GetPluginInfo(ctx, req, PluginName)
+}
+
+// GetPluginCapabilities implementation
+func (p *Plugin) GetPluginCapabilities(
+	ctx context.Context,
+	req *csi.GetPluginCapabilitiesRequest) (
+	*csi.GetPluginCapabilitiesResponse, error) {
+
+	return common.GetPluginCapabilities(ctx, req)
+}
